+++ test/bodyguard/policy_test.exs
@@ -1,5 +1,5 @@
 defmodule PolicyTest do
-  use ExUnit.Case, async: true
+  use ExUnit.Case, async: false
 
   setup do
     %{context: TestContext, user: %TestContext.User{}}